+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Journal des modifications
 - Module Admissions : Modification des sujets/corps de certains templates
 - Module Admissions : Changement du libellé du bouton à la dernière étape en fonction de l'état du dossier et de l'utilisateur connecté
 - Améliorations des doc et scripts d'install.
+- Améliorations de la page d'initialisation de compte utilisateur local.
 - [FIX] Utilisation du bon code structure pour les établissements (source_code).
 - [FIX] : Ajout de conditions pour éviter que InscriptionAssertion retourne une erreur lors de l'accès à un écran présentant des formations (ayant des sessions sans séances) ou d'une désinscription
 - [FIX] : Initialisation de compte : ajout du filtre 'StringTrim' sur les champs du formulaire.